Function additions
The system should completely function now. Additional buttons on the endgame allow scouters to mark a robot as having technical issues during a match. This will add a question to the pit scouting queue to visit this team later on. The end of the game will also create a scores question, recording scores and RP for each alliance. New reports will include scores and RP ranked for all teams. New pit questions ask about batteries and programming language. The team report now contains this information in addition to a breakdown of each match a team has been in.